+++ 30-enterprise-software-development-java/20-enterprise-framework/10-spring/code_example01/src/main/java/com/switchfully/spring/tax/TaxationDto.java
@@ -4,12 +4,12 @@ import java.time.LocalDate;
 
 public class TaxationDto {
     private final String fullName;
-    private final int brutoIncome;
+    private final double brutoIncome;
     private final double netIncome;
     private final double taxes;
     private LocalDate dueDate;
 
-    public TaxationDto(String fullName, int brutoIncome, double netIncome, double taxes, LocalDate dueDate) {
+    public TaxationDto(String fullName, double brutoIncome, double netIncome, double taxes, LocalDate dueDate) {
         this.fullName = fullName;
         this.brutoIncome = brutoIncome;
         this.netIncome = netIncome;
@@ -21,7 +21,7 @@ public class TaxationDto {
         return fullName;
     }
 
-    public int getBrutoIncome() {
+    public double getBrutoIncome() {
         return brutoIncome;
     }
