+++ 30-enterprise-software-development-java/20-enterprise-framework/10-spring/code_example01/src/main/java/com/switchfully/spring/tax/system/TaxSystem.java
@@ -4,5 +4,5 @@ import com.switchfully.spring.person.Person;
 import com.switchfully.spring.tax.Taxation;
 
 public interface TaxSystem {
-    Taxation calculateTaxFor(Person person, int income);
+    Taxation calculateTaxFor(Person person, double income);
 }
